+++ global.json
@@ -1,5 +1,5 @@
 {
   "sdk": {
-    "version": "5.0.400"
+    "version": "5.0.402"
   }
 }
