+++ src/Server/src/Abstractions/IIdentityServerService.cs
@@ -13,21 +13,4 @@ namespace IdOps
         Task<IEnumerable<IdentityServerKey>> GetKeysAsync(Guid id, CancellationToken cancellationToken);
         Task<Model.IdentityServer> SaveAsync(SaveIdentityServerRequest request, CancellationToken cancellationToken);
     }
-
-    public record SaveIdentityServerRequest(
-        string Name,
-        Guid EnvironmentId,
-        Guid GroupId,
-        string Url)
-    {
-        public Guid? Id { get; init; }
-    }
-
-    public record IdentityServerKey(string Kid, string Alg)
-    {
-        public string? Thumbprint { get; init; }
-        public string? SerialNumber { get; init; }
-        public string? Subject { get; init; }
-        public DateTime? ValidUntil { get; init; }
-    }
 }
