+++ src/Server/src/Abstractions/SaveIdentityServerGroupRequest.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+
+namespace IdOps
+{
+    public record SaveIdentityServerGroupRequest(
+        string Name,
+        ICollection<string> Tenants,
+        string Color)
+    {
+        public Guid? Id { get; init; }
+    }
+}
